# src/main/java/com/kleadingsolutions/expenseshare/dto/AddMemberRequest.java
package com.kleadingsolutions.expenseshare.dto;

import lombok.*;

import java.util.UUID;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class AddMemberRequest {
    private UUID userId;
}